requirements.txt
# Coordinate Conversions
navpy
pygeodesy
utm
geopy
pyproj

# Serial/ Hardware communication
pyserial
sparkfun-ublox-gps
pyubx2
crsf-parser
spidev

# Simulation/ RL
sailboat-gym
gymnasium==0.28.1
moviepy
msgpack-python==0.5.6
pyzmq==25.0.2
pydantic
multiexit

# Image processing/ Object Detection
opencv-python
Pillow
imageio-ffmpeg

# Telemetry/ Internet
requests

# documentation
mkdocs
mkdocs-material


# Misc
numpy
tqdm
pandas
docker
matplotlib
pydict2class
dataclass_wizard
click